pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aravalli"
version = "0.1.0"
description = "Aravalli Range Delineation Comparison Tool - OLD (FSI-like) vs NEW (20-Nov-2025)"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
authors = [
    {name = "Aravalli Research Team"}
]
keywords = ["aravalli", "geospatial", "delineation", "DEM", "GIS", "India"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "geopandas>=0.10.0",
    "shapely>=2.0.0",
    "pyproj>=3.2.0",
    "rasterio>=1.2.0",
    "rioxarray>=0.9.0",
    "xarray>=0.19.0",
    "elevation>=1.1.0",
    "osmnx>=1.2.0",
    "matplotlib>=3.4.0",
    "contextily>=1.2.0",
    "scikit-image>=0.18.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
    "rich>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
]

[project.scripts]
aravalli = "aravalli.cli:main"

[project.urls]
"Homepage" = "https://github.com/yourorg/aravalli"
"Bug Tracker" = "https://github.com/yourorg/aravalli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
